requirements.txt
numpy
meshio
matplotlib
opencv-python
tvb-gdist
eikonalfm
pyvista
scipy